% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtLHDs.R
\name{wtLHDs}
\alias{wtLHDs}
\title{Weighted Criteria-Based Latin Hypercube Designs (LHDs) for Any Numbers of Factors (>=2)}
\usage{
wtLHDs(
  levels,
  factors,
  w1,
  w2,
  w3,
  pop_size = 30,
  generations = 100,
  mut_prob = 0.05
)
}
\arguments{
\item{levels}{Range of levels,L is F<=L<=choose(F+2,2), where, F is number of factors.}

\item{factors}{Any number of factors, F >=2.}

\item{w1}{Weight of maximum absolute correlation. Between 0 to 1. So that w1+w2+w3=1.}

\item{w2}{Weight of Phi_p criterion. Between 0 to 1. So that w1+w2+w3=1.}

\item{w3}{Weight of Maxpro criterion. Between 0 to 1. So that w1+w2+w3=1.}

\item{pop_size}{Default population size is 30.}

\item{generations}{Default number of generations is 100.}

\item{mut_prob}{Mutation probability, by default it is 0.05.}
}
\value{
Generates Latin hypercube designs for a given number of factor-level combinations.
}
\description{
Weighted Criteria-Based Latin Hypercube Designs (LHDs) for Any Numbers of Factors (>=2)
}
\examples{
library(CompExpDes)
wtLHDs_prime(9,3,0.5,0.5,0)
}
